+++ CHANGELOG.md
@@ -3,9 +3,11 @@ All notable changes to the gtk-fortran project are documented in this file. The 
 
 ## [Development]
 
+### Added
+- `examples/bazaar.f90`: the About button credits the authors of that file. The call to gtk_about_dialog_set_license() is replaced by the more convenient gtk_about_dialog_set_license_type().
+
 ### Fixed
 - `g_application_run()` should be called with an array `[c_null_ptr]` as third argument instead of ``c_null_ptr`. Needed with the NAG Fortran compiler.
-
 
 ## [gtk-fortran 4.1] 2021-10-22
 The gtk-4-fortran library has been generated from GTK 4.4.0 and GLib 2.70.0 under Fedora 35.